% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_radar.R
\name{create_interactive_radar}
\alias{create_interactive_radar}
\title{Plot an interactive radar/spider-chart of the ESD data towards a given
reference, using the radarchart package}
\usage{
create_interactive_radar(ESD_data, reference, thresh = 0.5)
}
\arguments{
\item{ESD_data}{tidy version of the ESD data representation}

\item{reference}{char choose an option from: wug, gemeente, vlaanderen,
wug_gemeente, wug_provincie, wug_vlaanderen}

\item{threshold}{provide the value on which a red threshold line should be
provided}
}
\description{
Plot an interactive radar/spider-chart of the ESD data towards a given
reference, using the radarchart package
}

