% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_loss_bar.R
\name{create_loss_bar}
\alias{create_loss_bar}
\title{Plot an stacked graph of the relative percentages of landuse in each of the
spatial entities: WUG, municipality and province}
\usage{
create_loss_bar(lu_data_ha)
}
\arguments{
\item{lu_data_ha}{tidy version of the landuse data representation}
}
\value{
ggplot barplot
}
\description{
Plot an stacked graph of the relative percentages of landuse in each of the
spatial entities: WUG, municipality and province
}

