% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stacked_bar.R
\name{create_stacked_bar}
\alias{create_stacked_bar}
\title{Plot an stacked graph of the relative percentages of landuse in each of the
spatial entities: WUG, municipality and province}
\usage{
create_stacked_bar(lu_data, base.text.size = 16)
}
\arguments{
\item{lu_data}{tidy version of the landuse data representation}

\item{base.text.size}{int size of the textual graph elements}
}
\value{
ggplot barplot
}
\description{
Plot an stacked graph of the relative percentages of landuse in each of the
spatial entities: WUG, municipality and province
}

