% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_to_tidy.R
\name{excel_to_tidy_csv}
\alias{excel_to_tidy_csv}
\title{Title}
\usage{
excel_to_tidy_csv(xls_file, path)
}
\arguments{
\item{xls_file}{char path name of the excel file interested in}

\item{path}{path to put the files in}
}
\description{
Title
}

