% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_link_table_data.R
\name{extract_link_table}
\alias{extract_link_table}
\title{Extract the data table linking the identifiers of the WUG to their
corresponding municipality and province}
\usage{
extract_link_table(xls_file, sheetname)
}
\arguments{
\item{xls_file}{char path name of the excel file interested in}

\item{sheetname}{char sheet name}
}
\value{
tibble with the link information of WUG to municiaplity and province
}
\description{
Extract the data table linking the identifiers of the WUG to their
corresponding municipality and province
}

