% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_landuse_data.R
\name{extract_lu_data}
\alias{extract_lu_data}
\title{Extract all landuse data and combine into a single tidy data set}
\usage{
extract_lu_data(lu_sheets, xls_file, columns)
}
\arguments{
\item{lu_sheets}{list of char with the sheet names}

\item{xls_file}{char path name of the excel file interested in}

\item{columns}{list of char with the column names to use}
}
\value{
tibble with the tidy version of the lu-data
}
\description{
Extract all landuse data and combine into a single tidy data set
}

