% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sheet.R
\name{extract_sheet}
\alias{extract_sheet}
\title{Extract the data from a single ESD-data containing sheet of excel}
\usage{
extract_sheet(sheet_name, xls_file, columns)
}
\arguments{
\item{sheet_name}{char with the name of the ESD data containing sheet}

\item{xls_file}{char path name of the excel file interested in}
}
\value{
tibble with the tidy version of the esd-data sheet
}
\description{
Extract the data from a single ESD-data containing sheet of excel
}

