% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esd_data.R
\name{get_esd_data}
\alias{get_esd_data}
\title{Collect the ESD data for a given WUG from the tidy data representation}
\usage{
get_esd_data(esd_data, wug_link_data, id_wug)
}
\arguments{
\item{esd_data}{tidy representation of all ecosystem services data}

\item{wug_link_data}{tidy version of link table wug/municipality/province}

\item{id_wug}{wug identifier to get data from}
}
\description{
Collect the ESD data for a given WUG from the tidy data representation
}

