% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landuse_data.R
\name{get_landuse_data_ha}
\alias{get_landuse_data_ha}
\title{Collect the  LANDUSE data for a given WUG from the excel-sheet in ha:
TABLE:
 - LG_WUG_\% -> row id_wug, col B:S
 - LG_Gemeenten_\% -> row GEMEENTE, col B:S
 - LG_Provincies_\% -> row Provincie, col B:S}
\usage{
get_landuse_data_ha(xls_file, id_wug)
}
\arguments{
\item{xls_file}{excel file}

\item{id_wug}{char wug identifier}
}
\value{
data.frame
}
\description{
Collect the  LANDUSE data for a given WUG from the excel-sheet in ha:
TABLE:
 - LG_WUG_\% -> row id_wug, col B:S
 - LG_Gemeenten_\% -> row GEMEENTE, col B:S
 - LG_Provincies_\% -> row Provincie, col B:S
}

