% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_landuse_data.R
\name{get_landuse_data_pt}
\alias{get_landuse_data_pt}
\title{Collect the percentage landuse data for a given WUG from tidy data}
\usage{
get_landuse_data_pt(lu_data, wug_link_data, id_wug)
}
\arguments{
\item{lu_data}{tidy representation of all landuse data}

\item{wug_link_data}{tidy version of link table wug/municipality/province}

\item{id_wug}{wug identifier to get data from}
}
\description{
Collect the percentage landuse data for a given WUG from tidy data
}

