% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggradar.R
\name{ggradar}
\alias{ggradar}
\title{ggplot style radar chart plots}
\usage{
ggradar(plot.data, base.text.size = 14,
  axis.labels = colnames(plot.data)[-1], plot.extent.x.sf = 1,
  plot.extent.y.sf = 1.2, grid.line.width = 0.5,
  grid.line.linetype = "longdash", grid.line.colour = "grey",
  grid.line.alpha = 0.5, grid.label.ticks = c(0, 0.5, 1),
  grid.label.values = c("0\%", "50\%", "100\%"), grid.label.size = 7,
  grid.label.colour = "grey", grid.label.min = TRUE,
  axis.label.offset = 1.15, axis.label.size = 8,
  axis.label.colour = "grey", group.line.width = 1.5,
  group.point.size = 6, background.circle.colour = "grey",
  background.circle.transparency = 0.2, plot.legend = if (nrow(plot.data) >
  1) TRUE else FALSE, legend.text.size = grid.label.size,
  legend.text.colour = "grey", group.linecolors = c("#FF5A5F", "#FFB400",
  "#007A87", "#8CE071", "#7B0051", "#00D1C1", "#FFAA91", "#B4A76C", "#9CA299",
  "#565A5C", "#00A04B", "#E54C20"))
}
\arguments{
\item{plot.data}{data.frame with the first column the group names of the
individual rows. All the other columns define the individual radial axis}

\item{base.text.size}{int size of the text elements in the graph}

\item{axis.labels}{list default the titles of the second to last column names}

\item{plot.extent.x.sf}{float scaling factor of plot size in x direction}

\item{plot.extent.y.sf}{float scaling factor of plot size in y direction}

\item{grid.line.width}{width of the grid lines}

\item{grid.line.linetype}{linetype of the grid lines}

\item{grid.line.colour}{colour of the grid lines}

\item{grid.line.alpha}{transparency of text tabels used for the grid ticks}

\item{grid.label.ticks}{list of 3 values (min, mid, max), defining the  values
(according to data) to put the grid ticks}

\item{grid.label.values}{text tabels used for the grid ticks}

\item{grid.label.size}{size of text tabels used for the grid ticks}

\item{grid.label.colour}{colour of text tabels used for the grid ticks}

\item{grid.label.min}{bool plot the smallest value of the grid tick labels}

\item{axis.label.offset}{offset value to shift the labels from the grid cross sections}

\item{axis.label.size}{size of the axis labels}

\item{axis.label.colour}{colour of the axis labels}

\item{group.line.width}{line width of the data lines}

\item{group.point.size}{point size of the data}

\item{background.circle.colour}{colour of the background panel}

\item{background.circle.transparency}{transparency of the background panel}

\item{plot.legend}{boolean add a legend (TRUE) or not (FALSE)}

\item{legend.text.size}{size of the legend text}

\item{legend.text.colour}{colour of the legend text}

\item{group.linecolors}{colours of the lines, when the list of colors does
not match the  number of columns, the provided list of colors will be
replicated and colors will be reused}
}
\description{
ggplot style radar chart plots
}

