% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esd_data.R
\name{get_esd_data}
\alias{get_esd_data}
\title{Collect the ESD data for a given WUG from the excel-sheet:}
\usage{
get_esd_data(xls_file, id_wug)
}
\arguments{
\item{xls_file}{excel file}

\item{id_wug}{char wug identifier}
}
\value{
data.frame
}
\description{
TABLE:
 - ESD_Wug >- row id_wug, col F:U
 - ESD_Gemeente -> row GEMEENTE, col C:R
 - ESD_Vlaanderen -> row 'Vlaanderen', col C:R
 - ESD_Wug_Gemeente -> row GEMEENTE, col B:Q
 - ESD_Wug_Provincie -> row Provincie, col C:R
 - ESD_Wug_Vlaanderen -> row 'VlaamseWug' (all), col C:R
}

